refactor(frontend): remove global config/domains
Replace global parameter config with specific config for each view layer, eg. `config/view-layers/networks/domains`.
