+++ frontend/src/config/view-layers/networks/adaptation-options.json
