+++ frontend/src/config/view-layers/networks/adaptation-sector-layers.json
