+++ frontend/src/config/view-layers/networks/domains.ts
@@ -6,6 +6,28 @@ import {
 
 import adaptationOptions from './adaptation-options.json';
 
+interface TotalDamagesParams {
+  rcp: string;
+  epoch: number;
+}
+
+export const totalDamagesConfig: DataParamGroupConfig<TotalDamagesParams> = {
+  paramDefaults: {
+    rcp: 'baseline',
+    epoch: 2010,
+  },
+  paramDomains: {
+    rcp: ['baseline', '4.5', '8.5'],
+    epoch: [2010, 2050],
+  },
+  paramDependencies: {
+    rcp: ({ epoch }) => {
+      if (epoch === 2010) return ['baseline'];
+      if (epoch === 2050) return ['4.5', '8.5'];
+    },
+  },
+};
+
 export interface AdaptationOptionParams {
   sector: string;
   subsector: string;
